# src/main/resources/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Verdan Skoleadministrasjon API",
    "description": "REST API for Verdan skoleadministrasjonssystem. Støtter multi-institusjon med rollebasert tilgangskontroll (SUPER_ADMIN, INSTITUTION_ADMIN, TEACHER, STUDENT).\n\nAlle endepunkter (unntatt `/api/login`, `/api/auth/refresh` og `/api/health`) krever JWT Bearer-token i `Authorization`-headeren.\n\nAccess-tokens utløper etter 15 minutter. Bruk `/api/auth/refresh` for å fornye.",
    "version": "2.0.0",
    "contact": {
      "name": "Verdan Dev Team (Gruppe 2)"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Lokal utvikling"
    },
    {
      "url": "http://localhost:8081",
      "description": "Docker"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/admissions/portal": {
      "get": {
        "tags": ["Admissions"],
        "summary": "Get admission portal listings for the student's level",
        "description": "Returns all open admission periods visible to the current user. Filters by fromLevel (institution level the student is graduating from).",
        "parameters": [{ "name": "fromLevel", "in": "query", "schema": { "type": "string", "enum": ["UNGDOMSSKOLE", "VGS", "FAGSKOLE", "UNIVERSITET"] }, "description": "Filter by source institution level" }],
        "responses": { "200": { "description": "List of available admission periods with institution info" } }
      }
    },
...
    "/api/admissions/apply": {
      "post": {
        "tags": ["Admissions"],
        "summary": "Submit an application (STUDENT)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["periodId"], "properties": { "periodId": { "type": "integer" } } } } }
        },
        "responses": {
          "200": { "description": "Application submitted" },
          "400": { "description": "Validation error (not graduated, period closed, already applied)" }
        }
      }
    },
...
    "/api/admissions/applications/{id}/withdraw": {
      "put": {
        "tags": ["Admissions"],
        "summary": "Withdraw an application (STUDENT)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Application withdrawn" }, "400": { "description": "Cannot withdraw accepted application" } }
      }
    },
    "/api/admissions/applications/{id}/confirm": {
      "put": {
        "tags": ["Admissions"],
        "summary": "Confirm an accepted offer (STUDENT)",
        "description": "Student confirms they accept this offer. All other pending applications are automatically withdrawn.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Offer confirmed, other applications withdrawn" }, "400": { "description": "Application is not in ACCEPTED status" } }
      }
    },
    "/api/admissions/periods": {
      "get": {
        "tags": ["Admissions"],
        "summary": "List admission periods for this institution (ADMIN)",
        "responses": { "200": { "description": "List of admission periods" } }
      },
      "post": {
        "tags": ["Admissions"],
        "summary": "Create admission period (ADMIN)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["programId", "startDate", "endDate", "maxSeats"], "properties": { "programId": { "type": "integer" }, "startDate": { "type": "string", "format": "date" }, "endDate": { "type": "string", "format": "date" }, "maxSeats": { "type": "integer" } } } } }
        },
        "responses": { "201": { "description": "Period created" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/admissions/periods/{id}": {
      "put": {
        "tags": ["Admissions"],
        "summary": "Update admission period (ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "startDate": { "type": "string", "format": "date" }, "endDate": { "type": "string", "format": "date" }, "maxSeats": { "type": "integer" } } } } } },
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "tags": ["Admissions"],
        "summary": "Delete admission period (ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Deleted" }, "404": { "description": "Not found" } }
      }
    },
    "/api/admissions/periods/{id}/close": {
      "put": {
        "tags": ["Admissions"],
        "summary": "Close an admission period (ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Period closed" } }
      }
    },
    "/api/admissions/periods/{id}/reopen": {
      "put": {
        "tags": ["Admissions"],
        "summary": "Reopen a closed admission period with a new deadline (ADMIN)",
        "description": "Reopens a closed period. Requires a new future-dated end date.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "newEndDate": { "type": "string", "format": "date", "description": "New deadline (must be in the future)" } } } } }
        },
        "responses": { "200": { "description": "Period reopened with new deadline" }, "400": { "description": "Deadline must be in the future" } }
      }
    },
    "/api/admissions/periods/{id}/process": {
      "post": {
        "tags": ["Admissions"],
        "summary": "Process applications — accept/reject based on seats and requirements (ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Processing results with accepted/rejected counts" } }
      }
    },
    "/api/admissions/periods/{id}/enroll-accepted": {
      "post": {
        "tags": ["Admissions"],
        "summary": "Enroll confirmed students into the program (ADMIN)",
        "description": "Creates ProgramMember entries for all CONFIRMED applications. Handles cross-level transfers (e.g. ungdomsskole→VGS at VG1) and same-level transfers (e.g. VGS→VGS at VG3 for påbygg).",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Enrollment results with count" } }
      }
    },
    "/api/admissions/periods/{id}/overview": {
      "get": {
        "tags": ["Admissions"],
        "summary": "Get overview of a period's applications (ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Application overview with counts by status" } }
      }
    },
    "/api/admissions/periods/{id}/requirements": {
      "get": {
        "tags": ["Admissions"],
        "summary": "Get admission requirements for a period (ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "List of requirements" } }
      },
      "post": {
        "tags": ["Admissions"],
        "summary": "Set admission requirement for a period (ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "subjectCode": { "type": "string" }, "minGrade": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Requirement saved" } }
      }
    },
    "/api/admissions/periods/{id}/bulk-publish": {
      "post": {
        "tags": ["Admissions"],
        "summary": "Bulk publish decisions to all applicants (ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Decisions published" } }
      }
    }
  },
...
}